Move utils one level up
