+++ packages/components/src/utils/getPlaceHolderImage.js
