+++ packages/components/src/utils/platforms.js
